apps/whoami-llm/pyproject.toml
[project]
name = "whoami-llm"
version = "0.1.0"
description = "Ask who someone is based on their Velog posts (local RAG)."
requires-python = ">=3.10"

dependencies = [
    "typer>=0.12",
    "requests>=2.31",
    "feedparser>=6.0.11",
    "html2text>=2024.2.26",
    "tiktoken>=0.7",
    "sentence-transformers>=2.7.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.4",
    "tqdm>=4.66.0",
    "huggingface-hub>=0.25.0",
]

[project.scripts]
whoami-llm = "whoami_llm.cli:app"
